requirements.txt
faiss_cpu==1.12.0
fastapi==0.116.1
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.75
langchain_google_genai==2.1.10
protobuf==6.32.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
uvicorn==0.35.0
google-genai
python-multipart